requirements.txt
PyYAML~=6.0
requests~=2.29.0
setuptools~=65.5.1